Extract recipe screen components
